select option becomes richer by getting arrays.
What changes you have made?
changed in format of  return statements from string to array in save and query

Why this change is necessary?
this change is to use save and query functions data.

What is the impact of the change?
selectOption get array instead of strings
we can use save and query beverage data